refactor(remove logger.py): remove logging to mongodb
instead the logs will be exposed on a volume. Storing the logs
in the mongodb seems inappropriate. Also transformation of the
logs would be necessary to keep them readable.

Decide later, whether logs should be stored on another medium.
