+++ .env.dist
@@ -18,4 +18,3 @@ ODTP_RESULT=
 TRANSFER_INPUT_TO_OUTPUT=FALSE
 ODTP_SAVE_SNAPSHOT=FALSE
 ODTP_SAVE_IN_RESULT=TRUE
-ODTP_LOGS_IN_DB=FALSE
