+++ startup.sh
@@ -9,7 +9,6 @@ source "/odtp/odtp-component-client/src/shell/traceback.sh"
 
 odtp::print_info "setting up log files for the component"
 touch /odtp/odtp-logs/log.txt
-touch /odtp/odtp-logs/odtpLoggerDebugging.txt
 touch /odtp/odtp-logs/odtpS3UploadedDebugging.txt
 
 odtp::print_info "starting the odtp client app"
